W5D1/Q1/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
langchain==0.2.0
langchain-openai==0.1.25
langchain-community==0.2.0
langchain-text-splitters==0.2.0
openai==1.40.0
httpx==0.27.2
chromadb==0.4.21
sentence-transformers==2.6.1
ragas==0.1.0
pandas==2.2.2
pydantic==2.7.1
pypdf==4.2.0
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Additional dependencies
datasets==2.14.0
numpy==1.24.3
python-multipart==0.0.6 